requirements.txt
streamlit==1.32.0
pymongo==4.6.1
python-dotenv==1.0.1
PyMuPDF==1.23.26
google-generativeai==0.3.2
Pillow==10.2.0
pandas==2.2.1
python-dotenv==1.0.1